[!] Now signed with unsigned comparsion is treated as error, not warning [!] Fixed signed/unsigned comparsion

+++ sources/main_node_tool.c
@@ -163,7 +163,7 @@ int main(int argc, const char **argv)
   }
 
   size_t l_size = sizeof(s_opts) / sizeof(struct options);
-  for (int i = 0; i < l_size; i++) {
+  for (size_t i = 0; i < l_size; i++) {
       int argv_index = 1;
       if (argc >= argv_index && !strncmp(s_opts[i].cmd, argv[argv_index], strlen (argv[argv_index]) + 1)) {
           int match = 1;
